% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-sensitivity.R
\name{avignon}
\alias{avignon}
\alias{avignon_sib}
\alias{avignon_si2h}
\alias{avignon_sim}
\title{Avignon insulin-sensitivity indices}
\usage{
avignon_sib(insulin_0, glucose_0, weight)

avignon_si2h(insulin_120, glucose_120, weight)

avignon_sim(sib, si2h)
}
\arguments{
\item{insulin_0, insulin_120}{Fasting and 2 h insulin, uU/mL.}

\item{glucose_0, glucose_120}{Fasting and 2 h glucose, mmol/L.}

\item{weight}{Body weight, kg.}

\item{sib, si2h}{Basal and 2 h Avignon indices (as returned by
\code{avignon_sib()} / \code{avignon_si2h()}), for \code{avignon_sim()}.}
}
\value{
A numeric vector of Avignon sensitivity values.
}
\description{
Avignon's Si indices scale the reciprocal of the insulin-glucose product by a
fixed glucose distribution volume (150 mL per kg body weight). \code{avignon_sib()}
uses the fasting (basal) state, \code{avignon_si2h()} the 2 h state, and
\code{avignon_sim()} combines them.
}
\details{
\deqn{S_{i,b} = 10^8 / (I_0 \times G_0 \times 150\,wt)}
\deqn{S_{i,2h} = 10^8 / (I_{120} \times G_{120} \times 150\,wt)}
\deqn{S_{i,M} = (0.137 \times S_{i,b} + S_{i,2h}) / 2}

Insulin is in uU/mL, glucose in mmol/L, weight in kg.
}
\examples{
sib  <- avignon_sib(insulin_0 = 10, glucose_0 = 5.3, weight = 85)
si2h <- avignon_si2h(insulin_120 = 60, glucose_120 = 7.8, weight = 85)
avignon_sim(sib, si2h)

}
\references{
Avignon A et al. (1999) \emph{Int J Obes Relat Metab Disord} 23:512-517.
}
