% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretion.R
\name{bigtt_air}
\alias{bigtt_air}
\alias{bigtt_air_30_120}
\alias{bigtt_air_60_120}
\title{BIGTT acute insulin response}
\usage{
bigtt_air_30_120(
  insulin_0,
  insulin_30,
  insulin_120,
  glucose_0,
  glucose_30,
  glucose_120,
  bmi,
  male
)

bigtt_air_60_120(
  insulin_0,
  insulin_60,
  insulin_120,
  glucose_0,
  glucose_60,
  glucose_120,
  bmi,
  male
)
}
\arguments{
\item{insulin_0, insulin_30, insulin_60, insulin_120}{Insulin at 0/30/60/120
min, uU/mL.}

\item{glucose_0, glucose_30, glucose_60, glucose_120}{Glucose at 0/30/60/120
min, mmol/L.}

\item{bmi}{Body mass index, kg/m^2.}

\item{male}{Sex indicator: \code{1}/\code{TRUE} for men, \code{0}/\code{FALSE} for women.}
}
\value{
A numeric vector of estimated acute insulin response.
}
\description{
The BIGTT-AIR indices estimate the acute (first-phase) insulin response from
reduced OGTT sampling, sex, and BMI (Hansen et al. 2007). Two forms use
different intermediate timepoints (30 or 60 min). \strong{Insulin is in uU/mL} and
\strong{glucose in mmol/L}.
}
\details{
\deqn{AIR_{0,30,120} = \exp(8.20 + 0.00178 I_0 + 0.00168 I_{30} - 0.000383 I_{120} - 0.314 G_0 - 0.109 G_{30} + 0.0781 G_{120} + 0.180\,male + 0.032\,BMI)}
\deqn{AIR_{0,60,120} = \exp(8.19 + 0.00339 I_0 + 0.00152 I_{60} - 0.000959 I_{120} - 0.389 G_0 - 0.142 G_{60} + 0.164 G_{120} + 0.256\,male + 0.038\,BMI)}
}
\examples{
bigtt_air_30_120(insulin_0 = 10, insulin_30 = 80, insulin_120 = 60,
                 glucose_0 = 5.3, glucose_30 = 9, glucose_120 = 7.5,
                 bmi = 30, male = 1)

}
\references{
Hansen T et al. (2007) BIGTT. \emph{Diabetes Care} 30:257-262.
}
