% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasting-sensitivity.R
\name{fasting_sensitivity}
\alias{fasting_sensitivity}
\alias{raynaud}
\alias{homa_ir}
\alias{firi}
\alias{quicki}
\alias{insulin_glucose_ratio}
\alias{glucose_insulin_ratio}
\alias{bennett_isi}
\alias{mcauley}
\alias{ohkura}
\alias{homa_ad}
\alias{belfiore_basal}
\title{Fasting insulin-sensitivity / resistance indices}
\usage{
raynaud(insulin)

homa_ir(insulin, glucose)

firi(insulin, glucose)

quicki(insulin, glucose)

insulin_glucose_ratio(insulin, glucose)

glucose_insulin_ratio(insulin, glucose)

bennett_isi(insulin, glucose)

mcauley(insulin, trig)

ohkura(cpeptide, glucose)

homa_ad(insulin, glucose, adiponectin)

belfiore_basal(insulin, glucose, ref_insulin = 9.46, ref_glucose = 5.08)
}
\arguments{
\item{insulin}{Fasting insulin, uU/mL (= mU/L).}

\item{glucose}{Fasting glucose. mmol/L for all functions except \code{quicki()},
which expects mg/dL.}

\item{trig}{Fasting triglycerides, mmol/L.}

\item{cpeptide}{Fasting C-peptide, nmol/L.}

\item{adiponectin}{Fasting adiponectin, ug/mL.}

\item{ref_insulin, ref_glucose}{Reference (normal) fasting insulin (uU/mL) and
glucose (mmol/L) for the Belfiore normalisation. Defaults are the
Belfiore (1998) normal means, 9.46 uU/mL and 5.08 mmol/L.}
}
\value{
A numeric vector the length of the recycled inputs.
}
\description{
Surrogate indices of insulin sensitivity or resistance computed from a single
fasting blood draw. Each function is a direct transcription of the published
formula and is vectorised over its inputs.
}
\details{
\tabular{ll}{
\code{raynaud()}                 \tab \eqn{40 / I_0} \cr
\code{homa_ir()}                 \tab \eqn{(I_0 \times G_0) / 22.5} \cr
\code{firi()}                    \tab \eqn{(I_0 \times G_0) / 25} \cr
\code{quicki()}                  \tab \eqn{1 / (\log_{10} I_0 + \log_{10} G_0)} \cr
\code{insulin_glucose_ratio()}   \tab \eqn{I_0 / G_0} \cr
\code{glucose_insulin_ratio()}   \tab \eqn{G_0 / I_0} \cr
\code{bennett_isi()}             \tab \eqn{1 / (\ln I_0 \times \ln G_0)} \cr
\code{mcauley()}                 \tab \eqn{\exp(2.63 - 0.28 \ln I_0 - 0.31 \ln TG)} \cr
\code{ohkura()}                  \tab \eqn{20 / (CP_0 \times G_0)} \cr
\code{homa_ad()}                 \tab \eqn{(I_0 \times G_0) / (22.5 \times ADIPO)} \cr
\code{belfiore_basal()}          \tab \eqn{2 / ((I_0/I_{ref})(G_0/G_{ref}) + 1)}
}

\code{homa_ir()}, \code{firi()}, \code{insulin_glucose_ratio()}, \code{glucose_insulin_ratio()},
\code{bennett_isi()}, \code{homa_ad()} and \code{belfiore_basal()} take fasting glucose in
\strong{mmol/L}; \code{quicki()} takes fasting glucose in \strong{mg/dL}. \code{mcauley()} takes
triglycerides in \strong{mmol/L}. \code{ohkura()} takes fasting C-peptide in nmol/L and
glucose in mmol/L. Insulin is in uU/mL (= mU/L) throughout, adiponectin in
ug/mL. Higher \code{homa_ir()}, \code{firi()}, \code{insulin_glucose_ratio()}, \code{homa_ad()}
mean more resistance; the others increase with sensitivity. \code{1 / homa_ir()}
gives the reciprocal-HOMA sensitivity index.
}
\examples{
homa_ir(insulin = 10, glucose = 5.5)
quicki(insulin = 10, glucose = 99)          # glucose in mg/dL
mcauley(insulin = 10, trig = 1.3)

}
\references{
Matthews DR et al. (1985) Homeostasis model assessment. \emph{Diabetologia}
28:412-419.
Katz A et al. (2000) QUICKI. \emph{J Clin Endocrinol Metab} 85:2402-2410.
Raynaud E et al. (1999) \emph{Diabetes Care} 22:1003-1004.
McAuley KA et al. (2001) \emph{Diabetes Care} 24:460-464.
Ohkura T et al. (2013) \emph{Diabetol Metab Syndr} 5:43.
Belfiore F et al. (1998) \emph{Mol Genet Metab} 63:134-141.
}
