% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glucagon.R
\name{glucagon}
\alias{glucagon}
\alias{glucagon_insulin_ratio}
\alias{glucagon_suppression}
\title{Glucagon indices}
\usage{
glucagon_insulin_ratio(glucagon_0, insulin_0)

glucagon_suppression(glucagon_0, glucagon_30, percent = FALSE)
}
\arguments{
\item{glucagon_0, glucagon_30}{Fasting and 30 min glucagon, pmol/L.}

\item{insulin_0}{Fasting insulin, pmol/L.}

\item{percent}{If \code{TRUE}, \code{glucagon_suppression()} returns the percentage
suppression; if \code{FALSE} (default), the absolute change.}
}
\value{
A numeric vector.
}
\description{
Simple fasting and OGTT glucagon measures. \code{glucagon_insulin_ratio()} is the
molar fasting glucagon:insulin ratio (both in pmol/L). \code{glucagon_suppression()}
quantifies the fall in glucagon from fasting to 30 min, either as an absolute
change or as a percentage of the fasting value.
}
\details{
\deqn{glucagon:insulin = GCG_0 / I_0}
\deqn{suppression_{abs} = GCG_0 - GCG_{30}}
\deqn{suppression_{pct} = 100 (GCG_0 - GCG_{30}) / GCG_0}
}
\examples{
glucagon_insulin_ratio(glucagon_0 = 15, insulin_0 = 70)
glucagon_suppression(glucagon_0 = 15, glucagon_30 = 9)              # absolute
glucagon_suppression(glucagon_0 = 15, glucagon_30 = 9, percent = TRUE)

}
