% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretion.R
\name{secretion}
\alias{secretion}
\alias{homa_b}
\alias{igi}
\alias{cir}
\alias{kadowaki}
\title{Beta-cell secretion / insulinogenic indices}
\usage{
homa_b(insulin, glucose)

igi(insulin_0, insulin_t, glucose_0, glucose_t)

cir(insulin_t, glucose_t)

kadowaki(insulin_0, insulin_30, glucose_30)
}
\arguments{
\item{insulin, insulin_0, insulin_t, insulin_30}{Insulin, uU/mL. \code{insulin_0} is
fasting; \code{insulin_t} is the post-load value at the chosen time.}

\item{glucose}{Fasting glucose, mmol/L (for \code{homa_b()}).}

\item{glucose_0}{Fasting glucose, mmol/L (for \code{igi()}).}

\item{glucose_t}{Post-load glucose: mmol/L for \code{igi()}, mg/dL for \code{cir()}.}

\item{glucose_30}{30 min glucose, mg/dL (for \code{kadowaki()}).}
}
\value{
A numeric vector the length of the recycled inputs.
}
\description{
Surrogates of insulin secretion and beta-cell function from fasting bloods
and the OGTT.
}
\details{
\tabular{ll}{
\code{homa_b()}   \tab \eqn{20 I_0 / (G_0 - 3.5)} (I uU/mL, G mmol/L) \cr
\code{igi()}      \tab \eqn{(I_t - I_0) / (G_t - G_0)} (I uU/mL, G mmol/L) \cr
\code{cir()}      \tab \eqn{I_t / (G_t (G_t - 70))} if \eqn{G_t > 70} (G mg/dL) \cr
\code{kadowaki()} \tab \eqn{(I_{30} - I_0) / G_{30}} (I uU/mL, G mg/dL)
}

\code{igi()} (the insulinogenic index) and \code{cir()} (corrected insulin response)
are timepoint-agnostic: pass the post-load insulin/glucose at whichever time
you want (30 min is the classic insulinogenic index). \code{cir()} returns \code{NA}
when \eqn{G_t \le 70} mg/dL to avoid a non-physiological denominator.
}
\examples{
homa_b(insulin = 10, glucose = 5.3)
igi(insulin_0 = 10, insulin_t = 80, glucose_0 = 5.3, glucose_t = 8.9)
cir(insulin_t = 80, glucose_t = 160)

}
\references{
Matthews DR et al. (1985) HOMA. \emph{Diabetologia} 28:412-419.
Seltzer HS et al. (1967) insulinogenic index. \emph{J Clin Invest} 46:323-335.
Sluiter WJ et al. (1976) corrected insulin response. \emph{Diabetes} 25:245-249.
}
